Import NONMEM output for parameter tables

Reads all relevant inputs (ie, parameter estimates, bootstrap, shrinkage, etc.) using directly model run or with the help of a YAML file. The parameter name must be provided in the model file. Other fields are optional and can be provided either in (1) the model file or a (2) YAML file. Few fields needs to be inputted directly from user to complete the table:

name (compulsory): name of parameter (eg, CL, Vd, nCL, nVd, etc.)

label: description of parameter (eg, "Volume of distribution", "Apparent Clearance", etc.)

units: parameter unit (eg, L/h, 1/hr, etc)

trans: parameter transformation (ie, %, exp, ilogit, sqrt, CV%)

type: Typical Values, Between Subject Variability, Inter Occasion Variability, Residual Error or Covariates (ie respectively, Structural, IIV, IOV, RUV or CovariateEffect)

parframe2setup(
  run_dir,
  run_prefix,
  runno,
  bootstrap = NULL,
  run_dir.boot = NULL,
  runno.boot = NULL,
  conf.level = 0.95,
  min_suc = TRUE,
  read.boot = NULL,
  boot.obj = NULL,
  yaml.file = NULL,
  yaml.file.name = NULL
)

Arguments

run_dir
NONMEM model directory

run_prefix
NONMEM run number prefix

runno
NONMEM run number

bootstrap
flag for availability of bootstrap results

run_dir.boot
bootstrap results directory

runno.boot
bootstrap NONMEM run number, if changes from runno

conf.level
bootstrap results confidence interval. By default set to 95%

min_suc
filter bootstrap results on minimization successful. By default set to TRUE

read.boot
flag for reading directly the customized filtering bootstrap raw_results.csv

boot.obj
customized filtered bootstrap raw_results.csv data frame passed by user

yaml.file
flag for using the yaml.file option; all the non-compulsory fields e.g., label, trans, units, type will be read from yaml file

yaml.file.name
name of the yaml file

Value