Create html table from parframe

Uses the output from parframe to generate a parameter estimates table in HTML

pmxpartab(
  parframe,
  meta,
  columns = c(value = "Estimate", rse = "RSE%", ci95 = "95%CI", shrinkage =
    "Shrinkage"),
  sections = TRUE,
  section.labels = c(Structural = "Typical Values", CovariateEffect =
    "Covariate Effects", RUV = "Residual Error", IIV = "Between Subject Variability", IOV
    = "Inter-Occasion Variability"),
  show.fixed.to.zero = F,
  na = "n/a",
  digits = 3
)

Arguments

parframe
object from parframe

meta
meta information data frame from parframe2setup

columns
which information to display in the table

sections
flag for having separated sections displayed

section.labels
which sections will be displayed

show.fixed.to.zero
flag to show or not parameter fixed to zero. By default will be not displayed.

na
how NA will be displayed

digits
how many digits will be displayed